pyproject.toml
[tool.poetry]
name = "tests-2"
version = "0.1.0"
description = ""
authors = ["Artem Burenin <burenin@mindbox.cloud>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
